+++ site/dat/docs/Changelog.md
@@ -8,6 +8,8 @@
  - do not change CWD when running JMeter
  - add forgotten Gatling script template to python egg
  - fix PassFail condition flaw with few datapoints
+ - explicitly write default values in jmx2yaml
+ - recognize JMX variables in jmx2yaml
 
 ## 1.4.3 <sup>14 apr 2016</sup>
  - bump up version for jmeter plugins installation to 1.4.0